@@ -17,10 +17,10 @@
     "source-map-loader": "^4.0.2"
   },
   "dependencies": {
-    "@frontend-watch-dog/web-sdk": "workspace:^0.0.1",
+    "@frontend-watch-dog/web-sdk": "^0.0.1",
     "node-fetch": "^3.3.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.1",
+    "react-dom": "^18.0.1",
     "react-router-dom": "^6.21.2",
     "source-map": "^0.7.4",
     "web-vitals": "^3.5.0"
