@@ -3,13 +3,13 @@ import { BluBiuResponseCode, responseMsg } from './context.type';
 
 
 export interface BluBiuResponse<T>{
-  code: BluBiuResponseCode
-  message:string
-  data:T
+  code: BluBiuResponseCode;
+  message: string;
+  data: T;
 }
 
 export default {
-  result(code:BluBiuResponseCode, data?:any) {
+  result(code: BluBiuResponseCode, data?: any) {
     const ctx = this as Context;
     ctx.body = {
       code,
@@ -18,7 +18,7 @@ export default {
     };
   },
 
-  success(data?:any) {
+  success(data?: any) {
     this.result(BluBiuResponseCode.SUCCESS, data);
   },
 
