@@ -3,13 +3,13 @@ import kafka, { KafkaClient, Producer, KafkaClientOptions, ProduceRequest, Messa
 import { Application } from 'egg';
 import { Topics } from './type';
 export class Kafka {
-  private config:KafkaClientOptions;
+  private config: KafkaClientOptions;
   private app: Application;
   private client: KafkaClient;
   private producer: Producer;
-  private isInit:boolean;
+  private isInit: boolean;
 
-  constructor(config:KafkaClientOptions, app:Application) {
+  constructor(config: KafkaClientOptions, app: Application) {
     this.config = config;
     this.app = app;
     this.client = new kafka.KafkaClient(this.config);
@@ -19,17 +19,17 @@ export class Kafka {
       this.isInit = true;
       this.app.coreLogger.info('[kafka] the producer is ready.');
     });
-    this.producer.on('error', err => {
+    this.producer.on('error', (err) => {
       this.app.coreLogger.error(`[kafka] have error ${err}`);
     });
   }
 
-  consumer(topic:Topics, onMessage:(messge:Message)=>void) {
+  consumer(topic: Topics, onMessage: (messge: Message) => void) {
     this.client.createTopics([{
       topic,
       partitions: 1,
       replicationFactor: 1,
-    }], err => {
+    }], (err) => {
       if (err) {
         console.log('forEach', err);
         return;
@@ -44,22 +44,22 @@ export class Kafka {
           autoCommit: true,
         },
       );
-      consumer.on('error', err => {
+      consumer.on('error', (err) => {
         this.app.coreLogger.error(`[kafka] consumer have error ${err}`);
       });
-      consumer.on('message', (message:Message) => {
+      consumer.on('message', (message: Message) => {
         onMessage && onMessage(message);
       });
     });
   }
 
-  send(topic:Topics, messages:any) {
+  send(topic: Topics, messages: any) {
     if (this.isInit) {
-      const message:ProduceRequest = {
+      const message: ProduceRequest = {
         topic,
         messages: JSON.stringify(messages),
       };
-      this.producer.send([ message ], err => {
+      this.producer.send([ message ], (err) => {
         if (err) assert(err, '[kafka] producer send err. errmsg ${err}');
       });
     } else {
@@ -68,6 +68,6 @@ export class Kafka {
   }
 }
 
-export const useKafka = (app:Application) => {
+export const useKafka = (app: Application) => {
   app.kafka = new Kafka(app.config.kafka, app);
 };
