@@ -5,9 +5,9 @@ import { createIndexName } from '@/app/utils';
 
 export default class TrafficMysqlService extends Service {
 
-  private getMysqlTableName = (appId:string) => createIndexName(this.app.config.appIndexName.page_traffics, appId);
+  private getMysqlTableName = (appId: string) => createIndexName(this.app.config.appIndexName.page_traffics, appId);
 
-  private async getModel(appId:string):Promise<sequelize.ModelCtor<sequelize.Model<TrafficModelIn>>> {
+  private async getModel(appId: string): Promise<sequelize.ModelCtor<sequelize.Model<TrafficModelIn>>> {
     const tableName = this.getMysqlTableName(appId);
     const model = this.app.model.define(tableName, TrafficModel, {
       indexes: [
@@ -21,7 +21,7 @@ export default class TrafficMysqlService extends Service {
     return model;
   }
 
-  async createTrafficTable(appId:string) {
+  async createTrafficTable(appId: string) {
     const tableName = this.getMysqlTableName(appId);
     const model = this.app.model.define(tableName, TrafficModel, {
       indexes: [
@@ -35,14 +35,14 @@ export default class TrafficMysqlService extends Service {
     await model.sync();
   }
 
-  async insertData(data:TrafficModelIn) {
+  async insertData(data: TrafficModelIn) {
     const model = await this.getModel(data.appId);
     await model.create({
       ...data,
     });
   }
 
-  async getTrafficData(appId:string, { beginTime, endTime, type, pageUrl }) {
+  async getTrafficData(appId: string, { beginTime, endTime, type, pageUrl }) {
     const model = await this.getModel(appId);
     return await model.findAll({
       where: {
