@@ -2,24 +2,24 @@ import * as sequelize from 'sequelize';
 
 export interface TrafficModelIn {
   /** 应用AppId */
-  appId:string
+  appId: string;
   /** 统计类型 1 分时统计 2 按日统计 */
-  type: 1 | 2
+  type: 1 | 2;
   /** 访问Url路径 */
-  pageUrl:string
+  pageUrl: string;
   /** pv */
-  pageViews:number
+  pageViews: number;
   /** uv 访问数量 */
-  uniqueVisitors:number
+  uniqueVisitors: number;
   /** ip 访问数量 */
-  uniqueIPsCount:number
+  uniqueIPsCount: number;
   /** 统计时间 */
-  statisticsTime: Date
-  createdAt?: string
-  updatedAt?:string
+  statisticsTime: Date;
+  createdAt?: string;
+  updatedAt?: string;
 }
 
-export const TrafficModel:sequelize.ModelAttributes = {
+export const TrafficModel: sequelize.ModelAttributes = {
   id: { type: sequelize.INTEGER, primaryKey: true, autoIncrement: true },
   appId: {
     type: sequelize.STRING,
