@@ -1,9 +1,9 @@
 import { Service } from 'egg';
 import dayjs from 'dayjs';
 class RedisDayNewUsersService extends Service {
-  private getKeyName = (appId:string, dateString:string) => `${appId}-dayNewUsers-${dateString}`;
+  private getKeyName = (appId: string, dateString: string) => `${appId}-dayNewUsers-${dateString}`;
 
-  async analyseDayNewUsers(appId:string, userId:string) {
+  async analyseDayNewUsers(appId: string, userId: string) {
     const isHas = await this.app.redis.sismember(`${appId}-allUsers`, userId);
     if (!isHas) {
       const key = this.getKeyName(appId, dayjs().format('YYYY-MM-DD'));
@@ -12,11 +12,11 @@ class RedisDayNewUsersService extends Service {
     }
   }
 
-  async getOneDayNewUsers(appId:string, dateString?:string) {
+  async getOneDayNewUsers(appId: string, dateString?: string) {
     return await this.app.redis.get(this.getKeyName(appId, dayjs(dateString).format('YYYY-MM-DD'))) || 0;
   }
 
-  async getAllUsers(appId:string) {
+  async getAllUsers(appId: string) {
     return await this.app.redis.scard(`${appId}-allUsers`);
   }
 }
