@@ -1,4 +1,4 @@
-// This file is created by egg-ts-helper@1.35.1
+// This file is created by egg-ts-helper@2.1.0
 // Do not modify this file!!!!!!!!!
 /* eslint-disable */
 
